fix: ImagesAdapter 수정하여 불필요한 작업 제거
